[VKCI-219] [common-core] Make RDE.state type as a custom string type (#297)
* create custom string type for DefinedEntity.State


* fix issues in go vet


* rename the files


---------

(cherry picked from commit a660bf83e0184b06feffddbb3395dbbd304c3557)
